app.py
from flask import Flask, jsonify, request
import os
from dotenv import load_dotenv
import boto3
from datetime import datetime
import json

# Load environment variables
load_dotenv()

app = Flask(__name__)

# AWS Configuration
AWS_ACCESS_KEY = os.getenv('AWS_ACCESS_KEY')
AWS_SECRET_KEY = os.getenv('AWS_SECRET_KEY')
S3_BUCKET = os.getenv('S3_BUCKET', 'hospital-dr-bucket')
REGION = os.getenv('AWS_REGION', 'us-east-1')

# Initialize S3 client
s3_client = boto3.client(
    's3',
    aws_access_key_id=AWS_ACCESS_KEY,
    aws_secret_access_key=AWS_SECRET_KEY,
    region_name=REGION
)

# In-memory storage for demo (in production, use a database)
backup_status = {}

@app.route('/api/backup', methods=['POST'])
def create_backup():
    """Create a backup of hospital data"""
    try:
        data = request.json
        backup_id = f"backup_{datetime.utcnow().strftime('%Y%m%d_%H%M%S')}"
        
        # In a real application, this would save to S3
        # s3_client.put_object(
        #     Bucket=S3_BUCKET,
        #     Key=f"backups/{backup_id}.json",
        #     Body=json.dumps(data)
        # )
        
        backup_status[backup_id] = {
            'status': 'completed',
            'timestamp': datetime.utcnow().isoformat(),
            'data_size': len(json.dumps(data))
        }
        
        return jsonify({
            'message': 'Backup created successfully',
            'backup_id': backup_id,
            'status': backup_status[backup_id]
        }), 201
        
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/api/backup/<backup_id>', methods=['GET'])
def get_backup(backup_id):
    """Retrieve backup status"""
    if backup_id not in backup_status:
        return jsonify({'error': 'Backup not found'}), 404
        
    return jsonify({
        'backup_id': backup_id,
        'status': backup_status[backup_id]
    })

@app.route('/api/restore/<backup_id>', methods=['POST'])
def restore_backup(backup_id):
    """Restore data from backup"""
    try:
        # In a real application, this would retrieve from S3
        # response = s3_client.get_object(
        #     Bucket=S3_BUCKET,
        #     Key=f"backups/{backup_id}.json"
        # )
        # data = json.loads(response['Body'].read())
        
        return jsonify({
            'message': f'Successfully restored from backup {backup_id}',
            'status': 'completed',
            'timestamp': datetime.utcnow().isoformat()
        })
        
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/api/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({
        'status': 'healthy',
        'timestamp': datetime.utcnow().isoformat(),
        'backup_count': len(backup_status)
    })

if __name__ == '__main__':
    app.run(debug=True, host='0.0.0.0', port=5000)
